requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
minio==7.2.3
google-generativeai==0.7.0
edge-tts==6.1.9
pytest==8.0.0
pytest-asyncio==0.23.5
gTTS
mutagen
email-validator
